automation/requirements.txt
playwright==1.48.0
langgraph==0.2.45
langchain-core==0.3.15
python-dotenv==1.0.1
